% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bib_format.R
\name{get_bib_format}
\alias{get_bib_format}
\title{Path to .csl file}
\usage{
get_bib_format(name = NULL)
}
\arguments{
\item{name}{the .csl file name; if null returns a list of options}
}
\value{
an absolute path to the .bib file
}
\description{
Path to .csl file
}
