% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bib_path.R
\name{get_bib_path}
\alias{get_bib_path}
\title{Path to .bib file}
\usage{
get_bib_path(name = NULL)
}
\arguments{
\item{name}{the .bib file name; if null returns a list of options}
}
\value{
an absolute path to the .bib file
}
\description{
Path to .bib file
}
